pyproject.toml
[tool.poetry]
name = "LyuPo7_gendiff"
version = "1.0.0"
description = "Python project #2 with Hexlet: Gendiff"
authors = ["LyuPo <lport@frtk.ru>"]
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
repository = "https://github.com/LyuPo7/python-project-lvl2"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
pyYaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0.2" 
flake8 = "^3.8.3"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
